+++ CHANGELOG.md
@@ -10,4 +10,5 @@
 * API and pluggable mechanism for destroying old scoped beans on class reload
 * Improved reload support for session scoped beans (the old beans are now removed from the session)
 * Improved reload support for proxies in filters
+* Added support for reloading request scoped beans used inside session scoped beans
 * An improved proxy factory bean that specifies the target class